adds task to verify contract
